CaseStudy1/src/main/java/com/Assignment/CaseStudy1/Service/EmployeeService.java
package com.Assignment.CaseStudy1.Service;

import com.Assignment.CaseStudy1.Entity.Employee;
import com.Assignment.CaseStudy1.Repository.EmployeeRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class EmployeeService {

    @Autowired
    EmployeeRepository employeeRepository;


    public Employee saveEmployee(Employee employee) {
        return employeeRepository.save(employee);
    }

    public Optional<Employee> getEmployeeById(Long id) {
        return employeeRepository.findById(id);
    }

    public Optional<Employee> updateEmployeeById(Long id, Employee employee) {
        Optional<Employee> optionalEmployee = employeeRepository.findById(id);
        if (optionalEmployee.isPresent()) {
            Employee existingEmployee = optionalEmployee.get();
            existingEmployee.setName(employee.getName());
            existingEmployee.setEmailId(employee.getEmailId());
            employeeRepository.save(existingEmployee);
            return Optional.of(existingEmployee);
        } else {
            return Optional.empty();
        }
    }

}
